print all elements of a list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,8 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t nodes;
-
+	int nodes;
 	nodes = 0;
 
 	while (h)
